+++ src/app/_components/Pagebullets/index.tsx
@@ -1,21 +1,19 @@
 "use client";
 
-import { memo, useMemo } from "react";
+import { memo, useCallback, useMemo, useState } from "react";
 
 import styles from "@/app/_components/Pagebullets/index.module.scss";
 import Dot from "@/app/_components/svgs/Dot";
+import { useIntersectionObserver } from "@/app/_hooks/useIntersectionObserver";
 
 interface PageBulletsProps {
-  currentId: string;
+  defaultCurrentId: string;
   pageIds: string[];
-  handleDotClick: (id: string) => void;
 }
 
-const PageBullets = ({
-  currentId,
-  handleDotClick,
-  pageIds,
-}: PageBulletsProps) => {
+const PageBullets = ({ defaultCurrentId, pageIds }: PageBulletsProps) => {
+  const [currentId, setCurrentId] = useState(defaultCurrentId);
+
   const dotsWithDistance = useMemo(() => {
     const currentIndex = pageIds.indexOf(currentId);
     return pageIds.map((id, index) => ({
@@ -23,6 +21,15 @@ const PageBullets = ({
       id,
     }));
   }, [currentId, pageIds]);
+
+  useIntersectionObserver(pageIds, setCurrentId);
+
+  const handleDotClick = useCallback((id: string) => {
+    const el = document.getElementById(id);
+    if (el) {
+      el.scrollIntoView({ behavior: "smooth" });
+    }
+  }, []);
 
   return (
     <div className={styles.pageBullets}>
